completing requests function to run in parallel and getting correct data from responses
relates to #1
